fix: フロントエンドのESLintエラー3件を修正
- AuthContext: AuthProviderをAuthProvider.jsxに分離し、Fast Refreshを有効化
- CardModal: useEffectを削除しcardのkeyリマウントで初期化
- List: useEffectを削除しstartEditing()でtitleValをセット

Refs #51
